* origin/topic/timw/limit-string-and-container-lengths-in-logs:
  Add NEWS entry for field length limiting
  Tag truncated values with a flag, plus pack threading::Value better
  Remove length limiting on string fields for HTTP
  Make total_size counter a member in logging::Manager
  Remove using numeric_limits and just check for zero instead
  Expand the size of the log-size filters for x509
  Add options to filter at the stream level as well as globally
  Add a weird that gets emitted when strings/containers are over the limits
  Add metrics to track string and container fields limited by length
  Replace unused stream argument from RecordToLogRecord with WriterInfo
  Implement string- and container-length filtering at the log record level

+++ src/logging/Manager.cc
@@ -246,10 +246,17 @@ struct Manager::WriterInfo {
 
     std::shared_ptr<telemetry::Counter> total_writes;
     std::shared_ptr<telemetry::Counter> total_discarded_writes;
+    std::shared_ptr<telemetry::Counter> total_truncated_string_fields;
+    std::shared_ptr<telemetry::Counter> total_truncated_containers;
 
     WriterInfo(std::shared_ptr<telemetry::Counter> total_writes,
-               std::shared_ptr<telemetry::Counter> total_discarded_writes)
-        : total_writes(std::move(total_writes)), total_discarded_writes(std::move(total_discarded_writes)) {}
+               std::shared_ptr<telemetry::Counter> total_discarded_writes,
+               std::shared_ptr<telemetry::Counter> total_truncated_string_fields,
+               std::shared_ptr<telemetry::Counter> total_truncated_containers)
+        : total_writes(std::move(total_writes)),
+          total_discarded_writes(std::move(total_discarded_writes)),
+          total_truncated_string_fields(std::move(total_truncated_string_fields)),
+          total_truncated_containers(std::move(total_truncated_containers)) {}
 };
 
 struct Manager::Stream {
@@ -281,9 +288,12 @@ struct Manager::Stream {
     zeek_uint_t max_delay_queue_size = 1;
     bool evicting = false;
 
+    size_t max_field_string_bytes = 0;
+    size_t max_total_string_bytes = 0;
+    size_t max_field_container_elements = 0;
+    size_t max_total_container_elements = 0;
 
     ~Stream();
-
 
     const detail::DelayInfoPtr& GetDelayInfo(const detail::WriteContext& ctx);
 
@@ -491,7 +501,15 @@ Manager::Manager()
       total_log_writer_discarded_writes_family(
           telemetry_mgr->CounterFamily("zeek", "log-writer-discarded-writes",
                                        {"writer", "module", "stream", "filter-name", "path"},
-                                       "Total number of log writes discarded due to size limitations.")) {
+                                       "Total number of log writes discarded due to size limitations.")),
+      total_log_writer_truncated_string_fields_family(
+          telemetry_mgr->CounterFamily("zeek", "log-writer-truncated-string-fields",
+                                       {"writer", "module", "stream", "filter-name", "path"},
+                                       "Total number of logged string fields limited by length")),
+      total_log_writer_truncated_container_fields_family(
+          telemetry_mgr->CounterFamily("zeek", "log-writer-truncated-containers",
+                                       {"writer", "module", "stream", "filter-name", "path"},
+                                       "Total number of logged container fields limited by length")) {
     rotations_pending = 0;
 }
 
@@ -659,8 +677,12 @@ bool Manager::CreateStream(EnumVal* id, RecordVal* sval) {
     streams[idx]->columns = columns->Ref()->AsRecordType();
     streams[idx]->max_delay_interval = sval->GetField("max_delay_interval")->AsInterval();
     streams[idx]->max_delay_queue_size = sval->GetField("max_delay_queue_size")->AsCount();
-
     streams[idx]->enable_remote = id::find_val("Log::enable_remote_logging")->AsBool();
+
+    streams[idx]->max_field_string_bytes = sval->GetField("max_field_string_bytes")->AsCount();
+    streams[idx]->max_total_string_bytes = sval->GetField("max_total_string_bytes")->AsCount();
+    streams[idx]->max_field_container_elements = sval->GetField("max_field_container_elements")->AsCount();
+    streams[idx]->max_total_container_elements = sval->GetField("max_total_container_elements")->AsCount();
 
     DBG_LOG(DBG_LOGGING, "Created new logging stream '%s', raising event %s", streams[idx]->name.c_str(),
             event ? streams[idx]->event->Name() : "<none>");
@@ -1148,10 +1170,12 @@ bool Manager::WriteToFilters(const Manager::Stream* stream, zeek::RecordValPtr c
         assert(info);
 
         // Alright, can do the write now.
-        size_t total_size = 0;
-        auto rec = RecordToLogRecord(stream, filter, columns.get(), total_size);
+        total_record_size = 0;
+        total_string_bytes = 0;
+        total_container_elements = 0;
+        auto rec = RecordToLogRecord(w->second, filter, stream, columns.get());
 
-        if ( total_size > max_log_record_size ) {
+        if ( total_record_size > max_log_record_size ) {
             reporter->Weird("log_record_too_large", util::fmt("%s", stream->name.c_str()));
             w->second->total_discarded_writes->Inc();
             continue;
@@ -1388,7 +1412,21 @@ bool Manager::SetMaxDelayQueueSize(const EnumValPtr& id, zeek_uint_t queue_size)
     return true;
 }
 
-threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t& total_size) {
+static size_t calculate_allowed(size_t field_size, size_t max_field_size, size_t max_total_size, size_t current_total) {
+    size_t allowed = field_size;
+
+    if ( max_field_size != 0 && max_field_size < allowed ) {
+        allowed = max_field_size;
+    }
+    if ( max_total_size != 0 && (max_total_size - current_total) < allowed ) {
+        allowed = max_total_size - current_total;
+    }
+
+    return allowed;
+}
+
+
+threading::Value Manager::ValToLogVal(WriterInfo* info, const Stream* stream, std::optional<ZVal>& val, Type* ty) {
     if ( ! val )
         return {ty->Tag(), false};
 
@@ -1398,7 +1436,7 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
         case TYPE_BOOL:
         case TYPE_INT:
             lval.val.int_val = val->AsInt();
-            total_size += sizeof(lval.val.int_val);
+            total_record_size += sizeof(lval.val.int_val);
             break;
 
         case TYPE_ENUM: {
@@ -1417,14 +1455,14 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
                 lval.val.string_val.length = 0;
             }
 
-            total_size += lval.val.string_val.length;
+            total_record_size += lval.val.string_val.length;
 
             break;
         }
 
         case TYPE_COUNT:
             lval.val.uint_val = val->AsCount();
-            total_size += sizeof(lval.val.uint_val);
+            total_record_size += sizeof(lval.val.uint_val);
             break;
 
         case TYPE_PORT: {
@@ -1441,35 +1479,48 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
 
             lval.val.port_val.port = p & ~PORT_SPACE_MASK;
             lval.val.port_val.proto = pt;
-            total_size += lval.val.port_val.size();
+            total_record_size += lval.val.port_val.size();
             break;
         }
 
         case TYPE_SUBNET:
             val->AsSubNet()->Get().ConvertToThreadingValue(&lval.val.subnet_val);
-            total_size += lval.val.subnet_val.size();
+            total_record_size += lval.val.subnet_val.size();
             break;
 
         case TYPE_ADDR:
             val->AsAddr()->Get().ConvertToThreadingValue(&lval.val.addr_val);
-            total_size += lval.val.addr_val.size();
+            total_record_size += lval.val.addr_val.size();
             break;
 
         case TYPE_DOUBLE:
         case TYPE_TIME:
         case TYPE_INTERVAL:
             lval.val.double_val = val->AsDouble();
-            total_size += sizeof(lval.val.double_val);
+            total_record_size += sizeof(lval.val.double_val);
             break;
 
         case TYPE_STRING: {
             const String* s = val->AsString()->AsString();
-            char* buf = new char[s->Len()];
-            memcpy(buf, s->Bytes(), s->Len());
+            size_t allowed_bytes = calculate_allowed(static_cast<size_t>(s->Len()), stream->max_field_string_bytes,
+                                                     stream->max_total_string_bytes, total_string_bytes);
+
+            if ( allowed_bytes < static_cast<size_t>(s->Len()) ) {
+                lval.truncated = true;
+                reporter->Weird("log_string_field_truncated", util::fmt("%s", stream->name.c_str()));
+                info->total_truncated_string_fields->Inc();
+            }
+
+            if ( allowed_bytes == 0 )
+                return lval;
+
+            char* buf = new char[allowed_bytes];
+            memcpy(buf, s->Bytes(), allowed_bytes);
 
             lval.val.string_val.data = buf;
-            lval.val.string_val.length = s->Len();
-            total_size += lval.val.string_val.length;
+            lval.val.string_val.length = allowed_bytes;
+            total_record_size += allowed_bytes;
+            total_string_bytes += allowed_bytes;
             break;
         }
 
@@ -1479,7 +1530,7 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
             auto len = strlen(s);
             lval.val.string_val.data = util::copy_string(s, len);
             lval.val.string_val.length = len;
-            total_size += lval.val.string_val.length;
+            total_record_size += lval.val.string_val.length;
             break;
         }
 
@@ -1491,7 +1542,7 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
             auto len = strlen(s);
             lval.val.string_val.data = util::copy_string(s, len);
             lval.val.string_val.length = len;
-            total_size += lval.val.string_val.length;
+            total_record_size += lval.val.string_val.length;
             break;
         }
 
@@ -1508,32 +1559,61 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
             auto& set_t = tbl_t->GetIndexTypes()[0];
             bool is_managed = ZVal::IsManagedType(set_t);
 
-            zeek_int_t set_length = set->Length();
-            lval.val.set_val.vals = new threading::Value*[set_length];
+            size_t allowed_elements =
+                calculate_allowed(static_cast<size_t>(set->Length()), stream->max_field_container_elements,
+                                  stream->max_total_container_elements, total_container_elements);
 
-            for ( zeek_int_t i = 0; i < set_length && total_size < max_log_record_size; i++ ) {
+            if ( allowed_elements < static_cast<size_t>(set->Length()) ) {
+                lval.truncated = true;
+                reporter->Weird("log_container_field_truncated", util::fmt("%s", stream->name.c_str()));
+                info->total_truncated_containers->Inc();
+            }
+
+            if ( allowed_elements == 0 )
+                return lval;
+
+            lval.val.set_val.vals = new threading::Value*[allowed_elements];
+
+            for ( size_t i = 0; i < allowed_elements && total_record_size < max_log_record_size; i++ ) {
                 std::optional<ZVal> s_i = ZVal(set->Idx(i), set_t);
-                lval.val.set_val.vals[i] = new threading::Value(ValToLogVal(s_i, set_t.get(), total_size));
+                lval.val.set_val.vals[i] = new threading::Value(ValToLogVal(info, stream, s_i, set_t.get()));
                 if ( is_managed )
                     ZVal::DeleteManagedType(*s_i);
                 lval.val.set_val.size++;
             }
+
+            total_container_elements += lval.val.set_val.size;
 
             break;
         }
 
         case TYPE_VECTOR: {
             VectorVal* vec = val->AsVector();
-            zeek_int_t vec_length = vec->Size();
-            lval.val.vector_val.vals = new threading::Value*[vec_length];
+
+            size_t allowed_elements =
+                calculate_allowed(static_cast<size_t>(vec->Size()), stream->max_field_container_elements,
+                                  stream->max_total_container_elements, total_container_elements);
+
+            if ( allowed_elements < static_cast<size_t>(vec->Size()) ) {
+                lval.truncated = true;
+                reporter->Weird("log_container_field_truncated", util::fmt("%s", stream->name.c_str()));
+                info->total_truncated_containers->Inc();
+            }
+
+            if ( allowed_elements == 0 )
+                return lval;
+
+            lval.val.vector_val.vals = new threading::Value*[allowed_elements];
 
             auto& vv = vec->RawVec();
             auto& vt = vec->GetType()->Yield();
 
-            for ( zeek_int_t i = 0; i < vec_length && total_size < max_log_record_size; i++ ) {
-                lval.val.vector_val.vals[i] = new threading::Value(ValToLogVal(vv[i], vt.get(), total_size));
+            for ( size_t i = 0; i < allowed_elements && total_record_size < max_log_record_size; i++ ) {
+                lval.val.vector_val.vals[i] = new threading::Value(ValToLogVal(info, stream, vv[i], vt.get()));
                 lval.val.vector_val.size++;
             }
+
+            total_container_elements += lval.val.vector_val.size;
 
             break;
         }
@@ -1544,8 +1624,8 @@ threading::Value Manager::ValToLogVal(std::optional<ZVal>& val, Type* ty, size_t
     return lval;
 }
 
-detail::LogRecord Manager::RecordToLogRecord(const Stream* stream, Filter* filter, RecordVal* columns,
-                                             size_t& total_size) {
+detail::LogRecord Manager::RecordToLogRecord(WriterInfo* info, Filter* filter, const Stream* stream,
+                                             RecordVal* columns) {
     RecordValPtr ext_rec;
 
     if ( filter->num_ext_fields > 0 ) {
@@ -1595,9 +1675,9 @@ detail::LogRecord Manager::RecordToLogRecord(const Stream* stream, Filter* filte
         }
 
         if ( val )
-            vals.emplace_back(ValToLogVal(val, vt, total_size));
+            vals.emplace_back(ValToLogVal(info, stream, val, vt));
 
-        if ( total_size > max_log_record_size ) {
+        if ( total_record_size > max_log_record_size ) {
             return {};
         }
     }
@@ -1648,8 +1728,11 @@ WriterFrontend* Manager::CreateWriter(EnumVal* id, EnumVal* writer, WriterBacken
                                                        {"filter-name", instantiating_filter},
                                                        {"path", info->path}};
 
-    WriterInfo* winfo = new WriterInfo(zeek::log_mgr->total_log_writer_writes_family->GetOrAdd(labels),
-                                       zeek::log_mgr->total_log_writer_discarded_writes_family->GetOrAdd(labels));
+    WriterInfo* winfo =
+        new WriterInfo(zeek::log_mgr->total_log_writer_writes_family->GetOrAdd(labels),
+                       zeek::log_mgr->total_log_writer_discarded_writes_family->GetOrAdd(labels),
+                       zeek::log_mgr->total_log_writer_truncated_string_fields_family->GetOrAdd(labels),
+                       zeek::log_mgr->total_log_writer_truncated_container_fields_family->GetOrAdd(labels));
     winfo->type = writer->Ref()->AsEnumVal();
     winfo->writer = nullptr;
     winfo->open_time = run_state::network_time;
